refactor(clients): rename Base* → clean names (Client, SQLClient, HTTPClient)
- BaseClient → Client (root, in clients/client.py)
- BaseSQLClient → SQLClient
- AsyncBaseSQLClient → AsyncSQLClient
- BaseHTTPClient → HTTPClient

Backward-compat aliases preserved:
- clients/__init__.py: BaseClient = Client, ClientInterface = Client
- clients/sql.py: BaseSQLClient = SQLClient, AsyncBaseSQLClient = AsyncSQLClient
- clients/base.py: BaseHTTPClient = HTTPClient, BaseClient = Client
